Update __init__.py
NOTICE: Work in progress, some bugs may be present.

First go with Python, so may not be as elegant and I'm sure there were better ways to implement these changes. I needed some additional functionality and decided to work them into the settings so others may use them if they please.

New Settings Added:
gpioOptions: 0 or 1 - 0 Sets the GPIO port to be a PULL DOWN. 1 sets it to be a PULL UP. Default is 0 (Pull Down)
- I have a dual extruder with two limit switches as the sensor, because of this, I needed the ability to wire them in a way that one could be disabled (unplugged), but the other would still work. They're set as NORMALLY CLOSED so that when the filament is out, it completes the circuit.

pauseOptions: 0 or 1 - 0 Sets it to pause immediately. 1 sets it to pause at next Z-Layer change. Default is 0 (Pause Immediately)

To Do:
- Add ability to execute external Python script (for email notifications or whatever).
- Add ability to move away from print but maintain temperatures (execute GCode).